requirements.txt
google-api-python-client>=2.100.0
youtube-transcript-api>=1.0.0
anthropic>=0.40.0
python-telegram-bot>=21.0
schedule>=1.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
